Migrate FtCommunityPost and FtCommunityPoll to the composition API
