+++ src/renderer/components/FtCommunityPoll/FtCommunityPoll.css
