+++ src/renderer/components/FtCommunityPoll/FtCommunityPoll.vue
@@ -52,17 +52,44 @@
         v-if="!revealAnswer"
         class="option-text"
       >
-        <font-awesome-icon :icon="['fas', 'eye']" /> {{ $t('Channel.Community.Reveal Answers') }}
+        <FontAwesomeIcon :icon="['fas', 'eye']" /> {{ $t('Channel.Community.Reveal Answers') }}
       </div>
       <div
         v-else
         class="option-text"
       >
-        <font-awesome-icon :icon="['fas', 'eye-slash']" /> {{ $t('Channel.Community.Hide Answers') }}
+        <FontAwesomeIcon :icon="['fas', 'eye-slash']" /> {{ $t('Channel.Community.Hide Answers') }}
       </div>
     </div>
   </div>
 </template>
 
-<script src="./ft-community-poll.js" />
-<style scoped src="./ft-community-poll.css" />
+<script setup>
+import { computed, ref } from 'vue'
+import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
+
+import { formatNumber } from '../../helpers/utils'
+
+const props = defineProps({
+  data: {
+    type: Object,
+    required: true
+  }
+})
+
+const formattedVotes = computed(() => {
+  return formatNumber(props.data.totalVotes)
+})
+
+const revealAnswer = ref(false)
+
+/**
+ * Use smallest as it's resized to 125px anyways and they're usually all larger than that
+ * @param {{ height: number, width: number, url: string }[]} images
+ */
+function findSmallestPollImage(images) {
+  return images.reduce((prev, img) => (img.height < prev.height) ? img : prev, images[0]).url
+}
+</script>
+
+<style scoped src="./FtCommunityPoll.css" />
