+++ src/renderer/components/FtCommunityPost/FtCommunityPost.scss
